third_party/blink/web_tests/custom-elements/element-internals-aria-element-reflection.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<body>
  <script>
    class CustomElement extends HTMLElement {
      constructor() {
        super();
        this.internals = this.attachInternals();
      }
    }
    customElements.define('custom-element', CustomElement);
  </script>

  <custom-element id="custom1"></custom-element>
  <div id="activedescendant">Active descendant</div>
  <div id="controls">Controls</div>
  <div id="describedby">Described by</div>
  <div id="details">Details</div>
  <div id="errormessage">Error message</div>
  <div id="flowto">Flow to</div>
  <div id="labelledby">Labelled By</div>
  <div id="owns">Owns</div>

  <script>
    const element_properties = [
      ['ariaActiveDescendantElement', 'activedescendant']];
    const array_properties = [
      ['ariaControlsElements', 'controls'],
      ['ariaDescribedByElements', 'describedby'],
      ['ariaDetailsElements', 'details'],
      ['ariaErrorMessageElements', 'errormessage'],
      ['ariaFlowToElements', 'flowto'],
      ['ariaLabelledByElements', 'labelledby'],
      ['ariaOwnsElements', 'owns']];

    test(t => {
      const custom = document.getElementById('custom1');
      for (const [property, id] of element_properties) {
        assert_equals(custom.internals[property], null, `Could not retrieve ${property} from ElementInternals`);
      }
      for (const [property, id] of array_properties) {
        assert_equals(custom.internals[property], null, `Could not retrieve ${property} from ElementInternals`);
      }
    }, "Getting previously-unset ARIA element reflection properties on ElementInternals should return null, and not crash");

    test(t => {
      const custom = document.getElementById('custom1');
      for (const [property, id] of element_properties) {
        const related = document.getElementById(id);
        custom.internals[property] = related;
        assert_equals(custom.internals[property], related, `Could not retrieve ${property} from ElementInternals`);
      }
      for (const [property, id] of array_properties) {
        const related = document.getElementById(id);
        custom.internals[property] = [related];
        assert_array_equals(custom.internals[property], [related], `Could not retrieve ${property} from ElementInternals`);
      }
    }, "Getting ARIA element reflection properties on ElementInternals should return the value that was set, and not crash");

    test(t => {
      const custom = document.getElementById('custom1');
      for (const [property, id] of element_properties) {
        const related = document.getElementById(id);
        custom.internals[property] = related;
        assert_equals(custom.internals[property], related);

        custom.internals[property] = null;
        assert_equals(custom.internals[property], null, `Could not retrieve null ${property} from ElementInternals`);
      }
      for (const [property, id] of array_properties) {
        const related = document.getElementById(id);
        custom.internals[property] = [related];
        assert_array_equals(custom.internals[property], [related]);

        custom.internals[property] = [];
        assert_array_equals(custom.internals[property], []);

        custom.internals[property] = null;
        assert_equals(custom.internals[property], null, `Could not retrieve ${property} from ElementInternals`);
      }
    }, "Setting ARIA element reflection properties on ElementInternals to null should delete any previous value, and not crash");

  </script>
</body>
